backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Parsing
python-pptx==0.6.23
PyMuPDF==1.23.8
Pillow==10.1.0
opencv-python==4.8.1.78

# AI/ML Models
groq==0.11.0
openai==1.3.7
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
bert-score==0.3.13

# NLP & Text Analysis
spacy==3.7.2
textstat==0.7.3
yake==0.4.8
nltk==3.8.1

# PDF Report Generation
weasyprint==60.1
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Data Science
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
